move vitally logic to separate file from districts
